backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
PyMySQL==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
Flask-Limiter==3.5.0